requirements.txt
duckdb==0.8.1
pydantic==1.10.8
python-dotenv==1.0.0
Requests==2.28.1
openai==0.27.7
tiktoken==0.4.0
dashscope==1.3.1
pytest==7.4.0
pytest-cov==4.1.0
pytest-html==3.2.0
pytest-metadata==3.0.0